Add day20 solutions
Good day: 48 on first and 28 on second.

Did the brute force simulation for both before going back and
realizing I need only look at acceleration for the first.

Have an idea for an elegant solution to the second; iterate through
pairs and calculate time-of-collision stats; sort by time of incidence
and just check whether or not both members are alive.